Fix (v0.0.7):
BUG 1 — Environment= ignored (process.rs): The child now applies each configured KEY=VALUE via libc::setenv before execvp, so services get the configured environment merged over the inherited one instead of only the parent env.

BUG 2 — Event bus broken (event.rs): start_listener() now binds a UnixDatagram (matching emit()'s send_to). The remove_file was changed to only unlink the path when metadata.file_type().is_socket() (stale socket). API signatures unchanged.

BUG 3 — OnFailure restarts on clean exit (main.rs): maybe_restart_service now takes the ProcessStatus; OnFailure only restarts on Failed/Signaled, never on Exited (exit 0). Always still restarts on everything.

BUG 4 — check_process reports healthy services as dead (process.rs): Replaced waitpid (ECHILD for non-parents) with libc::kill(pid, 0). Returns Alive when the process exists (including EPERM), Unknown otherwise — preserving the one caller that relies on Unknown for dead-process cleanup.

BUG 5 — snapshot import broken (commands.rs): tmp_dir is now created with create_dir_all before tar xzf. Both the tar/rename and the directory-copy fallback check their results; success is only printed on success, with proper error messages and tmp-dir cleanup on failure.

BUG 6 — enable/disable has no effect (config.rs): load_all_services now reads /etc/cesar/enabled and filters services to those whose name appears there. If the dir is absent, all services load (previous behavior).

BUG 7 — config set writes outside any section (commands.rs): Missing keys are now appended under a [cesar] section header (the main section the parser understands) instead of dangling at EOF.

BUG 8 — socket destroy reports success even on failure (commands.rs): remove_file result is checked; failure prints an error.

BUG 9 — suspend --hybrid writes "freeze" (commands.rs): Hybrid now writes "disk" (same as hibernate) instead of "freeze" (s2idle).

BUG 10 — shutdown --reboot flag ignored (commands.rs): system shutdown --reboot now calls libc::reboot(libc::RB_AUTOBOOT); otherwise RB_POWER_OFF.

BUG 11 — service kill race (commands.rs): After sending SIGTERM, it polls check_process (kill 0) for up to 5s in 100ms steps before marking the service stopped. SIGKILL path unchanged.

BUG 12 — restart during shutdown (main.rs): maybe_restart_service returns early if SHUTDOWN_REQUESTED is set, and shutdown_graceful re-asserts the flag (the main loop had swapped it to false) so reaped children during the shutdown grace period are never auto-restarted.

BUG 13 — dead condition in boot loop (main.rs): Removed the always-true !svc.is_ready() term in both boot loops (service was just marked Starting, so it could never be "ready"); the dependency-satisfaction check remains intact, so behavior is unchanged.

MISSING — csl --tree exec path detection (bin/csl.rs): Now matches any Exec ... not found message and extracts the path by splitting on ' (second field), e.g. Exec binary '/system/bin/foo' not found! → /system/bin/foo. Also fixed the resulting needless-borrow clippy lint.

+++ src/health.rs
@@ -185,8 +185,8 @@ impl HealthReport {
         unsafe {
             let mut stat: libc::statvfs = std::mem::zeroed();
             if libc::statvfs(c"/".as_ptr().cast(), &mut stat) == 0 {
-                let total = stat.f_blocks as u64 * stat.f_frsize as u64;
-                let avail = stat.f_bavail as u64 * stat.f_frsize as u64;
+                let total = stat.f_blocks * stat.f_frsize;
+                let avail = stat.f_bavail * stat.f_frsize;
                 disk_used_pct = (total - avail).checked_mul(100).and_then(|v| v.checked_div(total)).unwrap_or(0);
                 disk_avail = format_bytes(avail);
             }
